Rename `typeerror`-rule to `type-error`
